pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = "Backend API for R3 Running App"
name = "r3-backend"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.13.1"
asyncpg = "^0.29.0"
fastapi = "^0.109.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.7.0"}
python = "^3.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
uvicorn = {extras = ["standard"], version = "^0.27.0"}

[tool.poetry.group.dev.dependencies]
httpx = "^0.26.0" # API 테스트 클라이언트
pytest = "^8.0.0" 

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
